Fixed movement when surrounded by bugs
Originally, given no choice but to encounter a bug, the bot would simply
pass its turn. This is problematic when multiple bugs are about to
collide with it. Instead, it now chooses the original closest path to an
item, while still hitting a bug.
TODO: Move when no choices can be made, if the enemy player is nearby
with a weapon