V1.3
Implemented bug direction. It was discovered that bugs cannot reverse
their direction, which significantly reduces the number of choices that
they can make. I have implemented 4 more lookup tables for bug movement,
each one corresponding to a different direction. Each table shows the
optimal path to take assuming that the bug cannot move backwards in the
direction it came from. As a result, the bot will be a lot less cautious
around bugs, because it only predicts legal bug movement. This will make
navigation far easier, especially when there are several bugs on the
board.