V1.4
-Disabled searching for swords when a sword is held.
-Reorganized most functions into individual files
-Modified the grid class to include all player information, so it can be
passed in a recursive search function in the future.
-Modified the findClosestItem function to take only a grid and player
id. From these, it can construct the board state for future moves. It
will still return a list of points to be played by the specified player.
TODO: Make this recursive, and make a variant on this function to
predict enemy player path. Currently, enemy just uses shortest path,
without accounting for bugs.
-Fixed a problem with paths.h, where the path from any point to itself
would be output as a 0, which represents 'up'. Now, this is replaced
with -1.