V1.5
-Significantly improved the underlying algorithm for evaluating moves.
-Enemy's movement will avoid bugs just like the player's.
-When the player does not have an optimal move to make, it will now
travel to the point on the board to optimize proximity to other spaces
relative to the opponent's position. A path from every point to the
enemy and to each other point is evaluated, and the point with the most
paths of shorter length than the enemy's is selected. This is all
hard-coded into another lookup table for performance.
-A new pathing function is added which allows the shortest path from a
player to any arbitrary point to be calculated, while avoiding bugs.
This is to help with pathing towards the optimal point on the board.